Rehauled search for bin data
Made a custom query that can return the row at which the level value
changes, use this to populate the table on the dashboard and will also
use it for the table on the silo page as well as the calendars.

The performance is really fast, generally using the custom query for
the hard work and then underscore for mappings.

Also turns out including region is is a bitch because of many-to-many
through relationships so I’m just leaving it out for now :/